Change layout generated by ubergen

+++ tools/@aws-cdk/ubergen/bin/ubergen.ts
@@ -6,7 +6,8 @@ import * as cfnspec from '@aws-cdk/cfnspec';
 import * as fs from 'fs-extra';
 import * as ts from 'typescript';
 
-const LIB_ROOT = path.resolve(process.cwd(), 'lib');
+const MONOPACKAGE_ROOT = process.cwd();
+const LIB_ROOT = process.cwd();
 const ROOT_PATH = findWorkspacePath();
 const UBER_PACKAGE_JSON_PATH = path.resolve(process.cwd(), 'package.json');
 
@@ -222,7 +223,10 @@ async function prepareSourceFiles(libraries: readonly LibraryReference[], packag
     console.log('\t 👩🏻‍🔬 \'excludeExperimentalModules\' enabled. Regenerating all experimental modules as L1s using cfn2ts...');
   }
 
-  await fs.remove(LIB_ROOT);
+  // Should not remove current directory if we're collecting into that
+  if (LIB_ROOT !== process.cwd()) {
+    await fs.remove(LIB_ROOT);
+  }
 
   const indexStatements = new Array<string>();
   for (const library of libraries) {
@@ -247,7 +251,7 @@ async function prepareSourceFiles(libraries: readonly LibraryReference[], packag
 async function combineRosettaFixtures(libraries: readonly LibraryReference[], uberPackageJson: PackageJson) {
   console.log('📝 Combining Rosetta fixtures...');
 
-  const uberRosettaDir = path.resolve(LIB_ROOT, '..', 'rosetta');
+  const uberRosettaDir = path.resolve(MONOPACKAGE_ROOT, 'rosetta');
   await fs.remove(uberRosettaDir);
   await fs.mkdir(uberRosettaDir);
 
@@ -325,12 +329,6 @@ async function transformPackage(
       },
       { spaces: 2 },
     );
-
-    await fs.writeFile(
-      path.resolve(LIB_ROOT, '..', `${library.shortName}.ts`),
-      `export * from './lib/${library.shortName}';\n`,
-      { encoding: 'utf8' },
-    );
   }
   return true;
 }
@@ -389,6 +387,7 @@ async function copyOrTransformFiles(from: string, to: string, libraries: readonl
       await fs.mkdirp(destination);
       return copyOrTransformFiles(source, destination, libraries, uberPackageJson);
     }
+
     if (name.endsWith('.ts')) {
       return fs.writeFile(
         destination,
